Add files via upload
The dataset has nine various features and was downloaded from Kaggle.
With the help of GridSearchCV, I have tried to find the best-fitted model for the dataset. Four different models were tested, and based on their accuracy, one model was selected.
The model takes a parameter and predicts whether the employee will leave the company or will be retained further.